+++ src/main/java/dk/kb/license/storage/AuditLogModuleStorage.java
@@ -2,7 +2,10 @@ package dk.kb.license.storage;
 
 import dk.kb.license.mapper.AuditLogEntryOutputDtoMapper;
 import dk.kb.license.model.v1.AuditLogEntryOutputDto;
+import dk.kb.license.model.v1.ObjectTypeEnumDto;
 import dk.kb.license.webservice.KBAuthorizationInterceptor;
+import dk.kb.util.webservice.exception.InvalidArgumentServiceException;
+
 import org.apache.cxf.jaxrs.utils.JAXRSUtils;
 import org.apache.cxf.message.Message;
 import org.keycloak.representations.AccessToken;
@@ -13,6 +16,7 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Class for handling the audit log entries.
@@ -23,33 +27,37 @@ public class AuditLogModuleStorage extends BaseModuleStorage {
     private static final AuditLogEntryOutputDtoMapper auditLogEntryOutputDtoMapper = new AuditLogEntryOutputDtoMapper();
 
     //AUDITLOG
-    private static final String AUDITLOG_TABLE = "AUDITLOG";
-    private static final String AUDITLOG_ID_COLUMN = "ID";
-    private static final String AUDITLOG_OBJECTID_COLUMN = "OBJECTID";
-    private static final String AUDITLOG_MODIFIEDTIME_COLUMN = "MODIFIEDTIME";
-    private static final String AUDITLOG_USERNAME_COLUMN = "USERNAME";
-    private static final String AUDITLOG_CHANGETYPE_COLUMN = "CHANGETYPE";
-    private static final String AUDITLOG_CHANGENAME_COLUMN = "CHANGENAME";
-    private static final String AUDITLOG_IDENTIFIER_COLUMN = "IDENTIFIER";
-    private static final String AUDITLOG_CHANGECOMMENT_COLUMN = "CHANGECOMMENT";
-    private static final String AUDITLOG_TEXTBEFORE_COLUMN = "TEXTBEFORE";
-    private static final String AUDITLOG_TEXTAFTER_COLUMN = "TEXTAFTER";
+    private static final String TABLE = "AUDITLOG";
+    private static final String ID_COLUMN = "ID";
+    private static final String OBJECTID_COLUMN = "OBJECTID";
+    private static final String MODIFIEDTIME_COLUMN = "MODIFIEDTIME";
+    private static final String USERNAME_COLUMN = "USERNAME";
+    private static final String CHANGETYPE_COLUMN = "CHANGETYPE";
+    private static final String CHANGENAME_COLUMN = "CHANGENAME";
+    private static final String IDENTIFIER_COLUMN = "IDENTIFIER";
+    private static final String CHANGECOMMENT_COLUMN = "CHANGECOMMENT";
+    private static final String TEXTBEFORE_COLUMN = "TEXTBEFORE";
+    private static final String TEXTAFTER_COLUMN = "TEXTAFTER";
+    
+                                
+    private final static String selectAuditLogOlderThanModifiedTimeQuery    = "SELECT * FROM " + TABLE + " WHERE " + MODIFIEDTIME_COLUMN + " < ? ORDER BY "+MODIFIEDTIME_COLUMN + " DESC LIMIT 100";
+    private final static String selectAuditLogOlderThanModifiedTimeByChangeNameQuery          = "SELECT * FROM " + TABLE + " WHERE " + MODIFIEDTIME_COLUMN + " < ? AND "+ CHANGENAME_COLUMN + " = ? ORDER BY " + MODIFIEDTIME_COLUMN + " DESC LIMIT 100";
+  
+    private final static String selectAuditLogQueryById = "SELECT * FROM " + TABLE + " WHERE " + ID_COLUMN + " = ? ";
+    private final static String selectAuditLogQueryByObjectId = "SELECT * FROM " + TABLE + " WHERE " + OBJECTID_COLUMN + " = ? " + " ORDER BY " + MODIFIEDTIME_COLUMN + " DESC";
 
-    private final static String selectAuditLogQueryById = "SELECT * FROM " + AUDITLOG_TABLE + " WHERE " + AUDITLOG_ID_COLUMN + " = ? ";
-    private final static String selectAuditLogQueryByObjectId = "SELECT * FROM " + AUDITLOG_TABLE + " WHERE " + AUDITLOG_OBJECTID_COLUMN + " = ? " + " ORDER BY " + AUDITLOG_MODIFIEDTIME_COLUMN + " DESC";
-
-    private final static String selectAllAuditLogQuery = "SELECT * FROM " + AUDITLOG_TABLE + " ORDER BY " + AUDITLOG_MODIFIEDTIME_COLUMN + " DESC";
-    private final static String persistAuditLog = "INSERT INTO " + AUDITLOG_TABLE + " (" +
-            AUDITLOG_ID_COLUMN + ", " +
-            AUDITLOG_OBJECTID_COLUMN + ", " +
-            AUDITLOG_MODIFIEDTIME_COLUMN + ", " +
-            AUDITLOG_USERNAME_COLUMN + ", " +
-            AUDITLOG_CHANGETYPE_COLUMN + ", " +
-            AUDITLOG_CHANGENAME_COLUMN + ", " +
-            AUDITLOG_IDENTIFIER_COLUMN + ", " +
-            AUDITLOG_CHANGECOMMENT_COLUMN + ", " +
-            AUDITLOG_TEXTBEFORE_COLUMN + ", " +
-            AUDITLOG_TEXTAFTER_COLUMN + ") " +
+    private final static String selectAllAuditLogQuery = "SELECT * FROM " + TABLE + " ORDER BY " + MODIFIEDTIME_COLUMN + " DESC";
+    private final static String persistAuditLog = "INSERT INTO " + TABLE + " (" +
+            ID_COLUMN + ", " +
+            OBJECTID_COLUMN + ", " +
+            MODIFIEDTIME_COLUMN + ", " +
+            USERNAME_COLUMN + ", " +
+            CHANGETYPE_COLUMN + ", " +
+            CHANGENAME_COLUMN + ", " +
+            IDENTIFIER_COLUMN + ", " +
+            CHANGECOMMENT_COLUMN + ", " +
+            TEXTBEFORE_COLUMN + ", " +
+            TEXTAFTER_COLUMN + ") " +
             "VALUES (?,?,?,?,?,?,?,?,?,?)"; // #|?|=10
 
     public AuditLogModuleStorage() throws SQLException {
@@ -78,6 +86,68 @@ public class AuditLogModuleStorage extends BaseModuleStorage {
         }
     }
 
+    /**
+     * Retrieves List of {@link AuditLogEntry} . Maximum elements is 100. Use modifiedTimeStart for pagination.
+     * AuditLogEntries in the list are sorted by modifiedtime desc, so latest will come first in the list.     
+     *      
+     * @param modifiedTimeStart Will extract AuditLogEntries with modifiedtime less than this value
+     * 
+     * @return List<AuditLogEntryOutputDto> with a maximum of 100 elements in the list. 
+     */
+   public List<AuditLogEntryOutputDto> getAuditLogOlderThanModifiedTimeListAll(Long modifiedTimeStart) throws IllegalArgumentException, SQLException {
+       log.debug("getAuditLogOlderThanModifiedTimeListAll called for modifiedTimeStart='{}', type='{}'", modifiedTimeStart);
+       
+       List<AuditLogEntryOutputDto> auditLogList = new ArrayList<AuditLogEntryOutputDto>();
+                       
+       try (PreparedStatement stmt = connection.prepareStatement(selectAuditLogOlderThanModifiedTimeQuery);) {
+           stmt.setLong(1, modifiedTimeStart);                     
+           ResultSet rs = stmt.executeQuery();
+           while (rs.next()) { // maximum one due to unique/primary key constraint              
+                AuditLogEntryOutputDto audit = auditLogEntryOutputDtoMapper.map(rs);
+                auditLogList.add(audit);
+           }       
+           return auditLogList;
+       } catch (SQLException e) {
+           log.error("SQL Exception in getAuditLogOlderThanModifiedTimeListAll " + e.getMessage());
+           throw e;
+       }
+   }
+   
+   /**
+    * Retrieves List of {@link AuditLogEntry} . Maximum elements is 100. Use modifiedTimeStart for pagination.
+    * AuditLogEntries in the list are sorted by modifiedtime desc, so latest will come first in the list.     
+    *      
+    * @param modifiedTimeStart Will extract AuditLogEntries with modifiedtime less than this value
+    * @param type Optional, list only AuditLogEntries of this type. Will list all types if type is null
+    * 
+    * @return List<AuditLogEntryOutputDto> with a maximum of 100 elements in the list. 
+    */
+  public List<AuditLogEntryOutputDto> getAuditLogOlderThanModifiedTimeListByType(Long modifiedTimeStart, ObjectTypeEnumDto changeName) throws IllegalArgumentException, SQLException {
+      if (changeName == null) {
+          throw new InvalidArgumentServiceException("ObjectTypeEnum must not be null");
+      }
+      
+      log.debug("getAuditLogOlderThanModifiedTimeListByType called for modifiedTimeStart='{}', changeName='{}'",modifiedTimeStart ,changeName);
+      
+      List<AuditLogEntryOutputDto> auditLogList = new ArrayList<AuditLogEntryOutputDto>();              
+      
+      try (PreparedStatement stmt = connection.prepareStatement(selectAuditLogOlderThanModifiedTimeByChangeNameQuery);) {
+          stmt.setLong(1, modifiedTimeStart);        
+          stmt.setString(2,changeName.getValue());                   
+          ResultSet rs = stmt.executeQuery();
+          while (rs.next()) { // maximum one due to unique/primary key constraint              
+               AuditLogEntryOutputDto audit = auditLogEntryOutputDtoMapper.map(rs);
+               auditLogList.add(audit);
+          }       
+          return auditLogList;
+      } catch (SQLException e) {
+          log.error("SQL Exception in getAuditLogOlderThanModifiedTimeListByType: " + e.getMessage());
+          throw e;
+      }
+  }
+
+   
+    
     /**
      * @param objectId The ID for the object extract audit log.
      * @return List of AuditLog objects. Will return empty list if objectId is not found.