+++ src/test/java/dk/kb/license/facade/LicenseModuleFacadeTest.java
@@ -11,6 +11,7 @@ import dk.kb.license.webservice.KBAuthorizationInterceptor;
 import org.apache.cxf.jaxrs.utils.JAXRSUtils;
 import org.apache.cxf.message.MessageImpl;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.keycloak.representations.AccessToken;
 import org.mockito.MockedStatic;
@@ -40,7 +41,27 @@ public class LicenseModuleFacadeTest extends UnitTestUtil {
         H2DbUtil.createEmptyH2DBFromDDL(URL, DRIVER, USERNAME, PASSWORD, List.of("ddl/licensemodule_create_h2_unittest.ddl", "ddl/audit_log_module_create_h2_unittest.ddl"));
         storage = new LicenseModuleStorageForUnitTest();
     }
-
+   
+    /*
+     * Delete all records between each unittest. The clearTableRecords is only called from here.
+     * The facade class is responsible for committing transactions. So clean up between unittests.
+     */
+    @BeforeEach
+    public void beforeEach() throws SQLException {
+        ArrayList<String> tables = new ArrayList<>();
+        tables.add("PRESENTATIONTYPES");
+        tables.add("GROUPTYPES");
+        tables.add("ATTRIBUTETYPES");
+        tables.add("LICENSE");
+        tables.add("ATTRIBUTEGROUP");
+        tables.add("ATTRIBUTE");
+        tables.add("VALUE_ORG");
+        tables.add("LICENSECONTENT");    
+        tables.add("PRESENTATION");
+        tables.add("AUDITLOG");    
+        storage.clearTableRecords(tables);
+    }
+    
     @Test
     public void testAuditLog() throws SQLException {
         HttpSession mockedSession = Mockito.mock(HttpSession.class);